refactor: create useRegistration hook and move the logic there
